rename *Model files to *CollisionModel
