+++ Sofa/Component/Collision/Geometry/CMakeLists.txt
@@ -7,6 +7,24 @@ set(HEADER_FILES
     ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/config.h.in
     ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/init.h
     ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/fwd.h
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/CubeCollisionModel.h
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/CylinderCollisionModel.h
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/CylinderCollisionModel.inl
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/LineCollisionModel.h
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/LineCollisionModel.inl
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/PointCollisionModel.h
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/PointCollisionModel.inl
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/RayCollisionModel.h
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/SphereCollisionModel.h
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/SphereCollisionModel.inl
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TetrahedronCollisionModel.h
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleCollisionModel.h
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleCollisionModel.inl
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleModelInRegularGrid.h
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleOctreeCollisionModel.h
+)
+
+set(DEPRECATED_HEADER_FILES
     ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/CubeModel.h
     ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/CylinderModel.h
     ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/CylinderModel.inl
@@ -20,27 +38,26 @@ set(HEADER_FILES
     ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TetrahedronModel.h
     ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleModel.h
     ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleModel.inl
-    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleModelInRegularGrid.h
     ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleOctreeModel.h
 )
 
 set(SOURCE_FILES
     ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/init.cpp
-    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/CubeModel.cpp
-    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/CylinderModel.cpp
-    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/LineModel.cpp
-    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/PointModel.cpp
-    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/RayModel.cpp
-    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/SphereModel.cpp
-    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TetrahedronModel.cpp
-    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleModel.cpp
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/CubeCollisionModel.cpp
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/CylinderCollisionModel.cpp
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/LineCollisionModel.cpp
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/PointCollisionModel.cpp
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/RayCollisionModel.cpp
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/SphereCollisionModel.cpp
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TetrahedronCollisionModel.cpp
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleCollisionModel.cpp
     ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleModelInRegularGrid.cpp
-    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleOctreeModel.cpp
+    ${SOFACOMPONENTCOLLISIONGEOMETRY_SOURCE_DIR}/TriangleOctreeCollisionModel.cpp
 )
 
 sofa_find_package(Sofa.Simulation.Core REQUIRED)
 
-add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${SOURCE_FILES} ${WRAPPER_FILES})
+add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${DEPRECATED_HEADER_FILES} ${SOURCE_FILES} ${WRAPPER_FILES})
 target_link_libraries(${PROJECT_NAME} PUBLIC Sofa.Simulation.Core)
 
 sofa_create_package_with_targets(
