+++ Sofa/Component/Collision/Geometry/src/sofa/component/collision/geometry/PointCollisionModel.cpp
