+++ Sofa/Component/Collision/Geometry/src/sofa/component/collision/geometry/RayModel.h
@@ -21,101 +21,6 @@
 ******************************************************************************/
 #pragma once
 #include <sofa/component/collision/geometry/config.h>
-#include <sofa/core/fwd.h>
-#include <sofa/core/CollisionModel.h>
-#include <sofa/defaulttype/VecTypes.h>
-#include <set>
+#include <sofa/component/collision/geometry/RayCollisionModel.h>
 
-namespace sofa::component::collision::response::contact
-{
-    class BaseRayContact;
-} // namespace sofa::component::collision::response::contact
-
-namespace sofa::component::collision::geometry
-{
-
-class RayCollisionModel;
-
-class SOFA_COMPONENT_COLLISION_GEOMETRY_API Ray : public core::TCollisionElementIterator<RayCollisionModel>
-{
-public:
-    Ray(RayCollisionModel* model, int index);
-
-    explicit Ray(const core::CollisionElementIterator& i);
-
-    const type::Vec3& origin() const;
-    const type::Vec3& direction() const;
-    SReal l() const;
-
-    void setOrigin(const type::Vec3& newOrigin);
-    void setDirection(const type::Vec3& newDirection);
-    void setL(SReal newL);
-};
-
-class SOFA_COMPONENT_COLLISION_GEOMETRY_API RayCollisionModel : public core::CollisionModel
-{
-public:
-    SOFA_CLASS(RayCollisionModel, core::CollisionModel);
-
-    typedef sofa::defaulttype::Vec3Types InDataTypes;
-    typedef sofa::defaulttype::Vec3Types DataTypes;
-    typedef Ray Element;
-    friend class Ray;
-protected:
-    RayCollisionModel(SReal defaultLength=1);
-public:
-    void init() override;
-
-    // -- CollisionModel interface
-    void resize(sofa::Size size) override;
-
-    void computeBoundingTree(int maxDepth) override;
-
-    void draw(const core::visual::VisualParams*, sofa::Index index) override;
-
-    core::behavior::MechanicalState<defaulttype::Vec3Types>* getMechanicalState() { return mstate; }
-    // ----------------------------
-    int addRay(const type::Vec3& origin, const type::Vec3& direction, SReal length);
-    Ray getRay(int index) { return Ray(this, index); }
-
-    int getNbRay() const { return size; }
-    void setNbRay(int n) { resize(n); }
-
-
-    void applyTranslation(const double dx,const double dy,const double dz);
-    virtual void addContact(response::contact::BaseRayContact* contact) { contacts.insert(contact); }
-    virtual void removeContact(response::contact::BaseRayContact* contact) { contacts.erase(contact); }
-
-    virtual const std::set<response::contact::BaseRayContact*> &getContacts() const { return contacts;}
-
-protected:
-    sofa::type::vector<SReal> length;
-    sofa::type::vector<type::Vec3> direction;
-
-    Data<SReal> d_defaultLength; ///< The default length for all rays in this collision model
-
-    std::set<response::contact::BaseRayContact*> contacts;
-    core::behavior::MechanicalState<defaulttype::Vec3Types>* mstate;
-
-};
-
-inline Ray::Ray(RayCollisionModel* model, int index)
-    : core::TCollisionElementIterator<RayCollisionModel>(model, index)
-{}
-
-inline Ray::Ray(const core::CollisionElementIterator& i)
-    : core::TCollisionElementIterator<RayCollisionModel>(static_cast<RayCollisionModel*>(i.getCollisionModel()), i.getIndex())
-{
-}
-
-inline void Ray::setDirection(const type::Vec3& newDirection)
-{
-    model->direction[index] = newDirection;
-}
-
-inline void Ray::setL(SReal newL)
-{
-    model->length[index] = newL;
-}
-
-} // namespace sofa::component::collision::geometry
+SOFA_HEADER_DEPRECATED("v25.12", "v26.06", "sofa/component/collision/geometry/RayCollisionModel.h")
