+++ Sofa/Component/Collision/Geometry/src/sofa/component/collision/geometry/SphereCollisionModel.cpp
